requirements.txt
# NearMeHQ Platform Dependencies
flask==3.0.0
gunicorn==21.2.0
stripe==7.0.0
twilio==9.0.0
supabase==2.3.0
resend==0.8.0
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
jinja2==3.1.3
markdown==3.5.2
